LayoutTests/http/tests/security/contentSecurityPolicy/default-src-inline-blocked-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "default-src 'self'".

CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "default-src 'self'".

This test passes if it doesn't alert fail.
